fixed file paths and exports to json now
